requirements.txt
fastapi 
uvicorn 
httpx 
itsdangerous
python-dotenv>=1.0